05_mem0_basics/pyproject.toml
[project]
name = "mem0-basics-demo"
version = "0.1.0"
description = "mem0 cloud memory quickstart using uv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mem0ai>=0.1.110",
]

[tool.uv]
package = false
